Fixed Login Redirect
The login redirect works fine on the login page and it works fine on the '/' page, but that portion was fixed in a weird way. Was not sure how else to fix the redirect from the login/register buttons.
I made admin page authorized only to administrators; it was not working correctly earlier.
Additional changes to layout page.